Ensure numeric status to show up normally
The previous commit caused numeric status to show up in bold fonts. This
commit ensures that colors and font weights are corrected in case a non
numeric status was shown (in bold, option `-o` for `set_color`).

To unify things, this commit also ensures correct colors and font
weights when indicating root shells. At the moment, it is the last
status that is being checked. However, additional status may be appended
in the future.